pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-event-git-tagging"
version = "0.0.1a1"
description = "Events for pythoneda-artifact/git-tagging"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact-event/git-tagging"
packages = [{include = "pythonedaartifacteventgittagging"}]

[tool.poetry.dependencies]
python = "^3.8"
pythoneda-base = "^0.0.1a15"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
